pyproject.toml
[project]
name = "agentic-appointment-assistant"
version = "0.1.0"
description = "Middleware-driven LangGraph appointment assistant with HITL"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"